starlight/starlight-api/src/main/java/com/baidu/cloud/starlight/api/model/ResultFuture.java
/*
 * Copyright (c) 2019 Baidu, Inc. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
 
package com.baidu.cloud.starlight.api.model;

import com.baidu.cloud.starlight.api.common.Constants;
import com.baidu.cloud.starlight.api.exception.StarlightRpcException;

import java.util.concurrent.CountDownLatch;
import java.util.concurrent.ExecutionException;
import java.util.concurrent.Future;
import java.util.concurrent.TimeUnit;
import java.util.concurrent.TimeoutException;

/**
 * Returned by an async call: Future Created by liuruisen on 2019/12/8.
 */
public class ResultFuture implements Future {

    private CountDownLatch countDownLatch = new CountDownLatch(1);

    private Response response;

    public void putResponse(Response response) {
        this.response = response;
        countDownLatch.countDown();
    }

    @Override
    public boolean cancel(boolean mayInterruptIfRunning) {
        return false;
    }

    @Override
    public boolean isCancelled() {
        return countDownLatch.getCount() == 0;
    }

    @Override
    public boolean isDone() {
        return response != null;
    }

    @Override
    public Object get() throws InterruptedException, ExecutionException {
        countDownLatch.await();
        return result();
    }

    @Override
    public Object get(long timeout, TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException {
        countDownLatch.await(timeout, unit);
        return result();
    }

    private Object result() {
        if (response.getStatus() != Constants.SUCCESS_CODE) {
            if (response.getException() instanceof RuntimeException) {
                throw (RuntimeException) response.getException();
            }
            throw new StarlightRpcException(response.getStatus(), response.getErrorMsg(), response.getException());
        }
        return response.getResult();
    }

}
